[PlutusCore] Split the normalizer
