+++ language-plutus-core/language-plutus-core.cabal
@@ -46,6 +46,7 @@ library
         Language.PlutusCore.Rename.Internal
         Language.PlutusCore.Rename
         Language.PlutusCore.Normalize
+        Language.PlutusCore.Normalize.Internal
         Language.PlutusCore.Pretty
         Language.PlutusCore.View
         Language.PlutusCore.Subst
