+++ language-plutus-core/prelude/PlutusPrelude.hs
@@ -19,6 +19,7 @@ module PlutusPrelude ( -- * Reëxports from base
                      , throw
                      , join
                      , (<=<)
+                     , (>=>)
                      , ($>)
                      , fromRight
                      , isRight
@@ -98,7 +99,7 @@ import           Control.Composition                     ((.*))
 import           Control.DeepSeq                         (NFData)
 import           Control.Exception                       (Exception, throw)
 import           Control.Lens
-import           Control.Monad                           (guard, join, (<=<))
+import           Control.Monad                           (guard, join, (<=<), (>=>))
 import           Data.Bifunctor                          (first, second)
 import           Data.Bool                               (bool)
 import qualified Data.ByteString.Lazy                    as BSL
