+++ language-plutus-core/test/Check/Spec.hs
@@ -6,8 +6,6 @@ import           Language.PlutusCore
 import qualified Language.PlutusCore.Check.Uniques          as Uniques
 import qualified Language.PlutusCore.Check.ValueRestriction as VR
 import           Language.PlutusCore.Generators.AST
-import           Language.PlutusCore.Quote
-import           PlutusPrelude
 
 import           Control.Monad.Except
 import           Data.Foldable                              (traverse_)
@@ -78,7 +76,7 @@ propRenameCheck :: Property
 propRenameCheck = property $ do
     prog <- forAll genProgram
     -- we didn't generate prog in Quote, so mark all the uniques as non-fresh
-    renamed <- runQuoteT $ (rename <=< through markNonFreshProgram) prog
+    renamed <- runQuoteT $ rename prog
     annotateShow renamed
     Hedgehog.evalExceptT $ checkUniques renamed
         where
